refactor(rdf-core): xsd_regex character classes and the x-flag stripper leave translate.rs
translate.rs had grown to 1101 lines, past the repo's 1000-line guidance, and
mixed three concerns: character-class construction, the `x`-flag whitespace
stripper, and the translator itself. This is a pure, behaviour-preserving
split with no semantic change.

- Move the character-class helpers (`push_hex_range`, `push_ranges`,
  `name_start_class_body`, `name_char_class_body`, `is_surrogate_range`, and
  the `\s`/`\S`/`\w`/`\W` bodies, lifted from inline literals into named consts
  carrying their XSD-production doc comments) into a new private `classes`
  module.
- Move `strip_x_flag_whitespace`, with its XPath F&O 3.1 §5.6.2 specification
  comment, into a new private `xflag` module.
- Move tests with the code they exercise.

Cross-module visibility is `pub(super)` only; the crate's public API is
unchanged and emitted regex source is byte-identical for every input. The
conformance corpus and harness are untouched. Step 1 of 2; the Scanner/Token
tokenizer refactor follows separately.
